# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Standup Comedy Connect API",
    "description": "API pour la plateforme de mise en relation entre humoristes et organisateurs d'événements de stand-up comedy",
    "version": "1.0.0",
    "contact": {
      "name": "Support API",
      "email": "support@standup-comedy-connect.com"
    }
  },
  "servers": [
    {
      "url": "https://standup-comedy-app.onrender.com/api",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Serveur de développement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Récupérer toutes les candidatures",
        "description": "Retourne les candidatures selon le rôle de l'utilisateur",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {"type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"]}
          },
          {
            "name": "eventId",
            "in": "query",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des candidatures",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {"type": "string"},
                      "event": {"type": "object"},
                      "comedian": {"type": "object"},
                      "status": {"type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"]},
                      "performanceDetails": {
                        "type": "object",
                        "properties": {
                          "duration": {"type": "number"},
                          "description": {"type": "string"},
                          "videoLink": {"type": "string"}
                        }
                      },
                      "message": {"type": "string"},
                      "organizerMessage": {"type": "string"},
                      "createdAt": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Créer une nouvelle candidature",
        "description": "Permet à un humoriste de postuler à un événement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["event", "performanceDetails"],
                "properties": {
                  "event": {"type": "string", "description": "ID de l'événement"},
                  "performanceDetails": {
                    "type": "object",
                    "required": ["duration", "description"],
                    "properties": {
                      "duration": {"type": "number", "minimum": 5, "maximum": 60},
                      "description": {"type": "string", "minLength": 10, "maxLength": 500},
                      "videoLink": {"type": "string", "format": "uri"}
                    }
                  },
                  "message": {"type": "string", "maxLength": 500}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Candidature créée avec succès"
          },
          "400": {
            "description": "Données invalides ou candidature déjà existante"
          }
        }
      }
    },
    "/applications/{applicationId}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Récupérer une candidature par ID",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la candidature"
          },
          "403": {
            "description": "Non autorisé à voir cette candidature"
          },
          "404": {
            "description": "Candidature non trouvée"
          }
        }
      },
      "put": {
        "tags": ["Applications"],
        "summary": "Mettre à jour le statut d'une candidature",
        "description": "Permet à l'organisateur d'accepter ou refuser une candidature",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {"type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"]},
                  "organizerMessage": {"type": "string", "maxLength": 500}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut mis à jour avec succès"
          },
          "403": {
            "description": "Non autorisé à modifier cette candidature"
          },
          "404": {
            "description": "Candidature non trouvée"
          }
        }
      },
      "delete": {
        "tags": ["Applications"],
        "summary": "Supprimer une candidature",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Candidature supprimée avec succès"
          },
          "403": {
            "description": "Non autorisé à supprimer cette candidature"
          },
          "404": {
            "description": "Candidature non trouvée"
          }
        }
      }
    },
    "/applications/{applicationId}/confirm": {
      "patch": {
        "tags": ["Applications"],
        "summary": "Confirmer la participation",
        "description": "Permet à un humoriste de confirmer sa participation après modification d'événement",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Participation confirmée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}